feat: literals and type promotion (#50)
This primarily adds two features:
1. Type promotion
2. Literals

**Type promotion**

Type promotion means that we automatically promote values where possible (e.g. in binary operations).
We adopt the JAX rules, which e.g. means that doing `i32 + f32` will promote the `i32` to an `f32`, and not an `f64` as one might expect. This is there to ensure that we don't up-promote to `f64`s too much, which slows down computation on accelerators. It is a trade-off between validity and speed though, as an `f32` can of course not hold all values an `i32` can hold.

**Literals**

It is now possible to use literals within jit-compiled programs.
For this, we introduced the `LiteralTensor` sub-class of `AbstractTensor` (which was previously `ShapedTensor`).
The `AbstractTensor` also got a field `@ambiguity` that indicates that we don't really know the type, which influences promotion behavior like in JAX. Within a graph, it's wrapped in a `GraphLiteral` (which already existed before, but wasn't used).

Other changes:
* Some API functions now call `to_abstract()` before accessing `shape()`, because they also need to handle literals and `shape(1L)` is not defined. We will document this in the [contributing guide](https://github.com/r-xla/anvil/pull/31) where it's added as a TODO.
* removed `dt_i32` etc; Use `dtype("i32")` now instead
* removed the spellchecker because it was annoying
* renamed `GraphBox@gval` -> `GraphBox@gnode`
* inference functions for graph builders no need to be manually defined not auto-convert between `stablehlo::FuncValue` and `anvil::ShapedTensor`, because we now require handling the ambiguity, which is not done by stablehlo.
* rename `nv_full` to `nv_fill`
* `GraphLiteral`s can have shapes `!= integer()`

Bug Fixes:
* `remove_dead_code` failed with higher-order primives, i.e., the sub-graphs were not checked for usage. We removed the pass now and let the XLA compiler handle this. In backward we now handle this by initializing 0-valued grads for ops that don't influence the output.